requirements.txt
fastapi
uvicorn
python-binance
python-dotenv
numpy
pandas
requests
streamlit
git